+++ test/test_history_size_unit.py
@@ -5,7 +5,7 @@ import os
 import unittest
 from unittest.mock import mock_open, patch
 
-from lshell.checkconfig import CheckConfig
+from lshell.config.runtime import CheckConfig
 from lshell.shellcmd import ShellCmd
 
 TOPDIR = f"{os.path.dirname(os.path.realpath(__file__))}/../"